+++ crates/easy-upnp/CHANGES.md
@@ -1,3 +1,17 @@
+# Changes in 0.2.0
+
+-   Add thiserror as dependency
+
+-   Use thiserror instead raw errors
+
+-   Properly propagate errors
+
+-   Use re-exported Ipv4Cidr
+
+-   Move error output to main, use Result in lib
+
+-   Update dependencies
+
 # Changes in 0.1.1
 
 -   Update dependencies for security fixes